Seeds, SFX and GFX
Added seeds for player to collect plus an on-screen seed counter. Made
powerup and seed collect sounds. Converted all SFX to OGG and MP3.
Adjusted the levels of the SFX. Created a new mute icon. Added a new
map including a non-colision layer for water and backdrop etc. Added
some code to map.js that horizontally tiles a background image to fit
the whole map. Changed font to increase readability. Created icons for
touch screen controls and hid a Tux penguin in the level ;)