Changed Jetpack Sound and Font.png
Added a new jetpack sound and changed how it’s played back. The sample
is triggered once when the jetpack is activated and then it loops over.
I changed the font.png to green so it contrasts the pink sky better, I
also removed the gradient to improve readability. Did a little bit more
level mixing on the SFX. I also changed the order in which the layers
are created in map.js, we can now overlay things on top of the backdrop
- See Tux Temple and the end of the level ;)